requirement.txt
Flask
python-docx
google-generativeai
python-dotenv
